+++ src/Action/ActionCancel.php
@@ -1,36 +1,35 @@
 <?php
 
 
-namespace TaskForceAction;
+namespace TaskForce\Action;
+
 
 
 /**
- * Class ActionDone
+ * Class ActionCancel
  * @package TaskForceAction
- * Класс наследуется от абстрактного, возвращает действие "Завершить"
+ * Класс наследуется от абстрактного, возвращает действие "Отменить"
  */
-class ActionDone extends AbstractSelectingAction
+class ActionCancel extends AbstractSelectingAction
 {
     /**
      * @var AbstractSelectingAction $checkingStatus
      * право на совершение действия
      */
     public $checkingStatus;
-
     /**
      * @return string|null
-     * метод по возврату названия действия при завершении
+     * метод по возврату названия действия при отмене
      */
     public function nameOfAction()
     {
         $checkingStatus = $this->checkingStatus;
         if ($this->$checkingStatus)
         {
-            return 'Завершить';
+            return 'Отменить';
         }
         return null;
     }
-
     /**
      * @return string|null
      */
@@ -39,17 +38,16 @@ class ActionDone extends AbstractSelectingAction
         $checkingStatus = $this->checkingStatus;
         if ($this->$checkingStatus)
         {
-            return 'action_done';
+            return 'action_cancel';
         }
         return null;
     }
-
     /**
      * @param $idPerformer
      * @param $idCustomer
      * @param $idUser
      * @return bool
-     * метод для проверки прав на совершение действия по завершению
+     * метод для проверки прав на совершение действия по отмене
      */
     public function checkingUserStatus($idPerformer, $idCustomer, $idUser)
     {
