+++ src/Action/ActionDone.php
@@ -1,36 +1,34 @@
 <?php
 
 
-namespace TaskForceAction;
+namespace TaskForce\Action;
 
 
 /**
- * Class ActionRefuse
+ * Class ActionDone
  * @package TaskForceAction
- * Класс наследуется от абстрактного, возвращает действие "Отказаться"
+ * Класс наследуется от абстрактного, возвращает действие "Завершить"
  */
-class ActionRefuse extends AbstractSelectingAction
+class ActionDone extends AbstractSelectingAction
 {
     /**
      * @var AbstractSelectingAction $checkingStatus
      * право на совершение действия
      */
     public $checkingStatus;
-
     /**
      * @return string|null
-     * метод по возврату названия действия при отказе
+     * метод по возврату названия действия при завершении
      */
     public function nameOfAction()
     {
         $checkingStatus = $this->checkingStatus;
         if ($this->$checkingStatus)
         {
-            return 'Отказаться';
+            return 'Завершить';
         }
         return null;
     }
-
     /**
      * @return string|null
      */
@@ -39,20 +37,19 @@ class ActionRefuse extends AbstractSelectingAction
         $checkingStatus = $this->checkingStatus;
         if ($this->$checkingStatus)
         {
-            return 'action_refuse';
+            return 'action_done';
         }
         return null;
     }
-
     /**
      * @param $idPerformer
      * @param $idCustomer
      * @param $idUser
      * @return bool
-     * метод для проверки прав на совершение действия по отказу
+     * метод для проверки прав на совершение действия по завершению
      */
     public function checkingUserStatus($idPerformer, $idCustomer, $idUser)
     {
-        return ($this->checkingStatus = ($idPerformer == $idUser));
+        return ($this->checkingStatus = ($idCustomer == $idUser));
     }
 }
