+++ src/Action/ActionRefuse.php
@@ -1,37 +1,34 @@
 <?php
 
 
-namespace TaskForceAction;
-
+namespace TaskForce\Action;
 
 
 /**
- * Class ActionCancel
+ * Class ActionRefuse
  * @package TaskForceAction
- * Класс наследуется от абстрактного, возвращает действие "Отменить"
+ * Класс наследуется от абстрактного, возвращает действие "Отказаться"
  */
-class ActionCancel extends AbstractSelectingAction
+class ActionRefuse extends AbstractSelectingAction
 {
     /**
      * @var AbstractSelectingAction $checkingStatus
      * право на совершение действия
      */
     public $checkingStatus;
-
     /**
      * @return string|null
-     * метод по возврату названия действия при отмене
+     * метод по возврату названия действия при отказе
      */
     public function nameOfAction()
     {
         $checkingStatus = $this->checkingStatus;
         if ($this->$checkingStatus)
         {
-            return 'Отменить';
+            return 'Отказаться';
         }
         return null;
     }
-
     /**
      * @return string|null
      */
@@ -40,20 +37,19 @@ class ActionCancel extends AbstractSelectingAction
         $checkingStatus = $this->checkingStatus;
         if ($this->$checkingStatus)
         {
-            return 'action_cancel';
+            return 'action_refuse';
         }
         return null;
     }
-
     /**
      * @param $idPerformer
      * @param $idCustomer
      * @param $idUser
      * @return bool
-     * метод для проверки прав на совершение действия по отмене
+     * метод для проверки прав на совершение действия по отказу
      */
     public function checkingUserStatus($idPerformer, $idCustomer, $idUser)
     {
-        return ($this->checkingStatus = ($idCustomer == $idUser));
+        return ($this->checkingStatus = ($idPerformer == $idUser));
     }
 }
