+++ src/test.php
@@ -1,7 +1,16 @@
 <?php
 require_once "../vendor/autoload.php";
 
-$idPerformer = 1;
-$idCustomer = 2;
-$strategy = new \TaskForce\Task($idPerformer, $idCustomer);
-assert($strategy->getNextStatus('action_cancel') == \TaskForce\Task::STATUS_CANCEL, 'cancel action');
+$idPerformer = 1; // id исполнителя задания
+$idCustomer = 2; // id заказчика задания
+$idUser = 2; // id текущего пользователя
+
+/**
+ * Создаём объект для определения списков действий и статусов, и выполнения базовой работы с ними
+ */
+$strategy = new \TaskForce\Task($idPerformer, $idCustomer, $idUser);
+$strategy->status = \TaskForce\Task::STATUS_NEW;
+
+// проверка
+
+var_dump($strategy->getNextStatus()); //возвращающение статуса в который перейдет задание
